refactor: InstagramPostResponse.java 패키지구조 수정
